refactor: helper function
